// SDKs/googleapis.com/dialogflow/v3/typescript/src/sdk/models/shared/googleclouddialogflowcxv3responsemessagemixedaudio.ts
import { Metadata, SpeakeasyBase } from "../../../internal/utils/utils";
import * as shared from "../shared";
import { GoogleCloudDialogflowCxV3ResponseMessageMixedAudioSegment } from "./googleclouddialogflowcxv3responsemessagemixedaudiosegment";


// GoogleCloudDialogflowCxV3ResponseMessageMixedAudio
/** 
 * Represents an audio message that is composed of both segments synthesized from the Dialogflow agent prompts and ones hosted externally at the specified URIs. The external URIs are specified via play_audio. This message is generated by Dialogflow only and not supposed to be defined by the user.
**/
export class GoogleCloudDialogflowCxV3ResponseMessageMixedAudio extends SpeakeasyBase {
  @Metadata({ data: "json, name=segments", elemType: shared.GoogleCloudDialogflowCxV3ResponseMessageMixedAudioSegment })
  segments?: GoogleCloudDialogflowCxV3ResponseMessageMixedAudioSegment[];
}